Add tests for auto depth and heading
These test are mostly for development purposes as they need to be run
when connected to a drone, hence the "connected_to_drone" marker on
the tests. They are handy because they allow expressing how a feature
is intended to be used, and ensure new functionality is tested
in a reproducible manner (i.e not by hand in iPython)  when developing.

A quirk related to running test using both a mocked and non-mocked tcp
client:
The test that run when connected to the drone using the non-mocked
Pioneer fixture need to be run before any test that use the mocked
tcp and udp clients. This is because as it is set up now the fixture
that mocks the tcp and upd client makes them stay mocked for the
remaining test for some reason (even when the fixture that mocks
them is marked as a function level fixture).