Add camera start and stop recording
A quirk with this is that for camera recording to be possible the
camera must be streaming somewhere. This could be a connected app, or
a gstreamer pipeline run from the SDK users computer. Because of this
the test for camera recording is skipped by default.

Starting and stopping recording feels very much like a service call, I
am not sure a property is the best interface for this functionality.