Make the intro.md page the home page of the documentation
intro.md is only a placeholder for now. But we did not want to have
the README.md be the home page of the documentation as it is more
developer centered. The easiest way to set the home page was to
configure nav.home in mkdocs, but this required having to manually
define the rest of the nav bar manually, this is okay since we want
more control of the order the documentation pages show up then having
them ordered alphabetically anyway.