Set  "About the Pioneer" as start page of docs
One issue with this, the relative links in this page works when
accessed from the home page, (i.e from http://127.0.0.1:8000/ if
served locally) but not when accessed from
http://127.0.0.1:8000/docs/about_the_pioneer/ but only when
about_the_pionee.md is in the docs dir . This seems to be
because the relative links in the page have a different start
directory when accesd through the start page or the "about the
Pioneer" page. A solution to this is moving the about_the_pioneer.md to
the repo root dir, but that is not nice either. Better solutions are
very welcome.