Include html documentation when publishing to PyPi
When connected to a drone you do not have access to the internet and
thus no access to the documentation, therefore we include the
documentation for local browsing in the PyPi package.
The "Update mkdocs config for local browsing" step  overwrites
pyproject.toml adding the parameter use_directory_urls = false to the
tools.portray.mkdocs configuration section, this makes browsing the
html documentation from your file system smoother by not having a
index.html file for each documentation section.

pyproject.toml is updated using sed because neither portray or poetry
supports updating the mkdocs configuration from the terminal and we do not want
to permanently update it, as it is not needed for the documentation
hosted on github pages.