Move tilt functionality to separate class
While the tilt function is closely related to the camera it doesn't
really make sense to have it as an extension of the Camera
class. Therefore I've separated it out into its own class, and rather
instantiate a tilt object on the camera class.

I've also made the get_tilt_angle function a @property since we've
used those fairly extensively elsewhere in the project. set_tilt_speed
can't be a property since we, as of this commit, don't have a way to
get the current speed from the drone and since it can be changed by the
drone at any point (due to end-stops, etc) it doesn't make sense to
keep a local state.