Make connection flow consistent for manual connect and auto connect
Now the behaviour is the same when doing
p = Pioneer(autoConnect=True)
or
p = Pioneer(autoConnect=False)
p.connect()
Previously the _tcp_client would be started before the _state_watcher
was started if autoConnect was set to True, and the order would be reveresed
if calling connect explicitly.
Also the state watcher thread is no longer started if the _tcp_client
could not connect, this avoids having the state watcher thread
start when no drone could be found.

Co-authored-by: Sindre Hansen <sindrehansen92@gmail.com>