Fix bug in connection flow for MacOS
In #74 we introduced a temporary UDP client used for waiting for the
drone to boot. Having multiple clients on the same ports works fine on
Windows and Linux, but turns out on Mac you have to set SO_REUSEPORT
and not SO_REUSEADDR. This was fixed in version 1.2.1 of
blueye.protocol so updating to this version fixes the issue.