Refactor how the TCP connection is handled
To allow for disconnecting and reconnecting to the drone the TCP
connection has to be released. Instead of trying to stop and start the
watchdog thread it's easier to simply stop everything and recreate the
object when it's time to reconnect.

Replacing the TcpClient object with a NoConnectionTcpClient object
makes for an easy way to ensure the user gets a more helpful error
message should they try to perform a command on a drone that's not
connected.

Moving "connection_established" variable to a property ensures there
are no breaking changes to the existing API.